+++ src/views/payment-sheet-views/card-view.js
@@ -215,7 +215,7 @@ CardView.prototype._generateHostedFieldsOptions = function () {
   if (!overrides) { return options; }
 
   if (overrides.fields) {
-    if (overrides.fields.cvv && overrides.fields.cvv.placeholder) {
+    if (overrides.fields.cvv && typeof overrides.fields.cvv.placeholder !== 'undefined') {
       this._hasCustomCVVPlaceholder = true;
     }